Inline setuputils.py to avoid issues with pip install from Homebrew
(similar problem and solution as applied to https://github.com/bluelabsio/db-facts)
